rename set_container_overrides to set_batch_overrides
